relezase11

+++ src/main/java/com/meli/desafiospringveterinaria/model/AnimalDomestico.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 //sem uso ate o momento
 public class AnimalDomestico extends Animal{
 
-    public AnimalDomestico(int numeroDoPaciente, String especie, String raca, String cor, LocalDate dataDeNascimento, String nome) {
+    public AnimalDomestico(long numeroDoPaciente, String especie, String raca, String cor, LocalDate dataDeNascimento, String nome) {
         super(numeroDoPaciente, especie, raca, cor, dataDeNascimento, nome);
     }
 }
